Polished UI button, Health Bar and what the player sees
Pause button, menu button and health bar have all been polished a bit more. Now the buttons from the in game UI to the main menu button have all got the same colour and stroke. On top of that we finally have a preview to "what the player can see"